Add CTDB event integration
CTDB can be configured to run event scripts at various times /
situations. This commit adds some initial plumbing to pass
these through to middlewared alerts and also subscribers to
ctdb events. The primary ones we are concerned with are:

INIT - add checks that middlewared is healthy and that the
ctdb shared volume is healthy before starting the ctdb process.
If this fails then ctdb will not start.

STARTUP - there are some services that we want to enable / disable
clusterwide. Add support for a special configuration file on
the ctdb_shared_volume that contains services that should always
be running when cluster is healthy. Services here will be restarted
when CTDB finishes starting.

MONITOR - CTDB monitors the services (same as STARTUP) at
configurable intervals and starts the service if it is stopped.
This also checks the status of the CTDB shared volume. Failure
of CTDB shared volume health check will cause CTDB on the node to
become UNHEALTHY. Failure of monitored service in this script
will not degrade CTDB status, but will generate alert / event.
Alert / event will only be triggered on state change based on
sentinel file.

STARTRECOVERY / RECOVERED - This sends events when recovery
operations are started and completed. Should be very infrequent
per documentation. Recovery master only.

IPREALLOCATED - triggered if public IP address changes between
nodes. This will generate an event from recovery master containing
new public IP info.

The actual event script called by CTDB will be located in
Samba repo, but use the CtdbEvent class defined here.

(cherry picked from commit 7848c95146f1886a98eacd3f7c53ef8997471e04)

+++ src/middlewared/middlewared/plugins/cluster_linux/ctdb_event_scripts.py
@@ -1,3 +1,5 @@
+import errno
+
 from pathlib import Path
 from enum import Enum
 
@@ -15,6 +17,7 @@ class CtdbAllowedEventScriptsEnum(Enum):
     Many ctdb event scripts exist so we only want these to be exposed, for now.
     """
     INTERFACE = '10.interface.script'
+    MIDDLEWARED = '80.truenas_middlewared.script'
 
 
 class CtdbEventScriptsService(Service):
@@ -33,6 +36,7 @@ class CtdbEventScriptsService(Service):
         # nodes in the cluster
         try:
             await self.middleware.call('ctdb.event.scripts.enable', '10.interface.script')
+            await self.middleware.call('ctdb.event.scripts.enable', '80.truenas_middlewared.script')
         except Exception:
             self.logger.error('Failed to initialize ctdb event scripts', exc_info=True)
 
@@ -48,7 +52,7 @@ class CtdbEventScriptsService(Service):
         """Return the ctdb event scripts that are disabled"""
         return [
             i.name for i in USR_DIR.iterdir() if i.is_file() and (
-                i.name in self.middleware.call_sync('ctdb.event.scripts.enabled')
+                i.name not in self.middleware.call_sync('ctdb.event.scripts.enabled')
             )
         ]
 
@@ -59,6 +63,14 @@ class CtdbEventScriptsService(Service):
 
         symlink_targ = ETC_DIR.joinpath(script)
         symlink_dest = USR_DIR.joinpath(script)
+        if not symlink_dest.exists():
+            raise CallError(
+                f'{symlink_dest}: event script does not exist. These scripts are part '
+                'of the TrueNAS distribution and so their absence may indicate a '
+                'broken or modified base installation. There are no steps to fix this in an '
+                'automated manner. Reinstall of operating system may be required.',
+                errno.ENOENT
+            )
         try:
             symlink_targ.unlink(missing_ok=True)  # make sure we remove a non-symlink if it exists
             symlink_targ.symlink_to(symlink_dest)
